start implementing workloadgroup based server
