+++ components/heimdall/cmd/server/main.go
@@ -15,234 +15,64 @@
 package main
 
 import (
-	"context"
-	"encoding/json"
-	"errors"
-	"fmt"
-	"net/http"
-	"os"
-
-	"github.com/gin-gonic/gin"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/client-go/kubernetes/scheme"
-	klog "k8s.io/klog/v2"
-	"sigs.k8s.io/controller-runtime/pkg/client"
+	istionetworkingv1beta1 "istio.io/client-go/pkg/apis/networking/v1beta1"
+	"k8s.io/apimachinery/pkg/runtime"
+	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
+	"k8s.io/klog/v2"
+	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client/config"
-	clientconfig "sigs.k8s.io/controller-runtime/pkg/client/config"
+	"sigs.k8s.io/controller-runtime/pkg/manager"
 
 	cluster_registry "github.com/cisco-open/cluster-registry-controller/api/v1alpha1"
-	istio_ca "github.com/cisco-open/nasp/pkg/ca/istio"
-	"github.com/cisco-open/nasp/pkg/environment"
+	"github.com/cisco-open/nasp/components/heimdall/pkg/controllers"
+	"github.com/cisco-open/nasp/components/heimdall/pkg/server"
+)
+
+var (
+	scheme = runtime.NewScheme()
 )
 
 func init() {
-	_ = cluster_registry.AddToScheme(scheme.Scheme)
-}
+	_ = clientgoscheme.AddToScheme(scheme)
 
-var podNamespace = os.Getenv("POD_NAMESPACE")
-var clusterID = os.Getenv("NASP_CLUSTER_ID")
-var istioVersion = os.Getenv("NASP_ISTIO_VERSION")
-var istioRevision = os.Getenv("NASP_ISTIO_REVISION")
-
-var ErrClientNotFound = errors.New("client not found in database")
-var ErrClusterIDNotFound = errors.New("clusterID not found")
-var ErrClientOrClientSecretInvalid = errors.New("invalid ClientID or ClientSecret")
-var ClientDatabaseConfigMap = types.NamespacedName{Namespace: podNamespace, Name: "heimdall-client-database"}
-
-type ConfigRequest struct {
-	ClientID     string `binding:"required"`
-	ClientSecret string `binding:"required"`
-	Version      string
-}
-
-type Client struct {
-	ClientID           string
-	ClientSecret       string
-	WorkloadName       string
-	PodNamespace       string
-	Network            string
-	MeshID             string
-	ServiceName        string
-	Version            string
-	ServiceAccountName string
-}
-
-type ClientDatabase interface {
-	Lookup(ClientID string) (*Client, error)
-}
-
-type ConfigMapClientDatabase struct {
-	c client.Client
-}
-
-func NewConfigMapClientDatabase() (ClientDatabase, error) {
-	kubeconfig, err := config.GetConfig()
-	if err != nil {
-		return nil, err
-	}
-
-	c, err := client.New(kubeconfig, client.Options{})
-	if err != nil {
-		return nil, err
-	}
-
-	// Sanity check
-	var configMap corev1.ConfigMap
-	err = c.Get(context.Background(), ClientDatabaseConfigMap, &configMap)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ConfigMapClientDatabase{c: c}, nil
-}
-
-func (db *ConfigMapClientDatabase) Lookup(clientID string) (*Client, error) {
-	var configMap corev1.ConfigMap
-	err := db.c.Get(context.Background(), ClientDatabaseConfigMap, &configMap)
-	if err != nil {
-		return nil, err
-	}
-
-	if clientData, ok := configMap.Data[clientID]; ok {
-		var client Client
-		err = json.Unmarshal([]byte(clientData), &client)
-		if err != nil {
-			return nil, err
-		}
-		return &client, nil
-	}
-
-	return nil, ErrClientNotFound
+	_ = cluster_registry.AddToScheme(scheme)
+	_ = istionetworkingv1beta1.AddToScheme(scheme)
 }
 
 func main() {
+	kubeconfig, err := config.GetConfig()
+	if err != nil {
+		panic(err)
+	}
+
 	logger := klog.TODO()
 
-	clientDb, err := NewConfigMapClientDatabase()
+	mgr, err := manager.New(kubeconfig, manager.Options{
+		Scheme:             scheme,
+		MetricsBindAddress: "0",
+	})
 	if err != nil {
 		panic(err)
 	}
 
-	if clusterID == "" {
-		clusterID, err = getClusterIDFromRegistry()
-		if err != nil {
+	if err := (&controllers.IstioWorkloadGroupReconciler{
+		Client: mgr.GetClient(),
+		Config: mgr.GetConfig(),
+		Scheme: mgr.GetScheme(),
+		Logger: logger,
+	}).SetupWithManager(mgr); err != nil {
+		panic(err)
+	}
+
+	ctx := ctrl.SetupSignalHandler()
+
+	go func() {
+		if err := mgr.Start(ctx); err != nil {
 			panic(err)
 		}
-	}
+	}()
 
-	r := gin.Default()
-	r.GET("/", func(c *gin.Context) {
-	})
-	r.POST("/config", func(c *gin.Context) {
-		var configRequest ConfigRequest
-		if err := c.ShouldBindJSON(&configRequest); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-			return
-		}
-
-		log := logger.WithValues("clientID", configRequest.ClientID)
-		log.Info("client requesting config")
-
-		client, err := clientDb.Lookup(configRequest.ClientID)
-		if err != nil {
-			log.Error(nil, "could not find client id")
-			c.JSON(http.StatusUnauthorized, gin.H{"error": ErrClientOrClientSecretInvalid.Error()})
-			return
-		}
-
-		if client.ClientSecret != configRequest.ClientSecret {
-			log.Error(nil, "client secret mismatch")
-			c.JSON(http.StatusUnauthorized, gin.H{"error": ErrClientOrClientSecretInvalid.Error()})
-			return
-		}
-
-		client.Version = configRequest.Version
-		if client.ServiceAccountName == "" {
-			client.ServiceAccountName = "default"
-		}
-
-		var e IstioCAClientConfigAndEnvironment
-
-		e.CAClientConfig, err = istio_ca.GetIstioCAClientConfigWithKubeConfig(clusterID, istioRevision, nil, &types.NamespacedName{
-			Name:      client.ServiceAccountName,
-			Namespace: client.PodNamespace,
-		})
-		if err != nil {
-			_ = c.AbortWithError(500, err)
-			return
-		}
-
-		clientIP := c.ClientIP()
-		if clientIP == "::1" {
-			clientIP = "127.0.0.1"
-		}
-
-		e.Environment = environment.IstioEnvironment{
-			Type:              "sidecar",
-			PodName:           client.WorkloadName + "-" + configRequest.ClientID,
-			PodNamespace:      client.PodNamespace,
-			PodOwner:          fmt.Sprintf("kubernetes://apis/v1/namespaces/%s/pods/%s", client.PodNamespace, client.WorkloadName),
-			PodServiceAccount: client.ServiceAccountName,
-			WorkloadName:      client.WorkloadName,
-			AppContainers:     nil,
-			InstanceIPs:       []string{clientIP},
-			Labels: map[string]string{
-				"security.istio.io/tlsMode":           "istio",
-				"service.istio.io/canonical-revision": "latest",
-				"istio.io/rev":                        e.CAClientConfig.Revision,
-				"topology.istio.io/network":           e.Environment.Network,
-				"k8s-app":                             client.WorkloadName,
-				"service.istio.io/canonical-name":     client.ServiceName,
-				"app":                                 client.WorkloadName,
-				"version":                             client.Version,
-			},
-			PlatformMetadata: nil,
-			Network:          client.Network,
-			SearchDomains:    []string{"svc.cluster.local", "cluster.local"},
-			ClusterID:        e.CAClientConfig.ClusterID,
-			DNSDomain:        "cluster.local",
-			MeshID:           client.MeshID,
-			IstioCAAddress:   e.CAClientConfig.CAEndpoint,
-			IstioVersion:     istioVersion,
-			IstioRevision:    istioRevision,
-		}
-
-		c.JSON(http.StatusOK, e)
-	})
-
-	if err := r.Run(); err != nil {
+	if err := server.New(logger).Run(ctx); err != nil {
 		panic(err)
 	}
-}
-
-func getClusterIDFromRegistry() (string, error) {
-	config, err := clientconfig.GetConfig()
-	if err != nil {
-		return "", err
-	}
-
-	k8sClient, err := client.New(config, client.Options{})
-	if err != nil {
-		return "", err
-	}
-
-	var clusters cluster_registry.ClusterList
-	err = k8sClient.List(context.Background(), &clusters)
-	if err != nil {
-		return "", err
-	}
-
-	for _, cluster := range clusters.Items {
-		if cluster.Status.Type == cluster_registry.ClusterTypeLocal {
-			return cluster.Name, nil
-		}
-	}
-
-	return "", ErrClusterIDNotFound
-}
-
-type IstioCAClientConfigAndEnvironment struct {
-	CAClientConfig istio_ca.IstioCAClientConfig
-	Environment    environment.IstioEnvironment
 }
