+++ components/heimdall/deploy/charts/heimdall/templates/serviceaccount.yaml
@@ -145,6 +145,12 @@ rules:
       - update
       - delete
       - patch
+  - apiGroups: [""]
+    resources: ["secrets"]
+    verbs: ["get", "watch", "list", "create", "delete", "update"]
+  - apiGroups: ["networking.istio.io"]
+    resources: ["workloadgroups"]
+    verbs: ["get", "watch", "list"]
 ---
 apiVersion: rbac.authorization.k8s.io/v1
 kind: ClusterRoleBinding
